+++ controller/dataplane/owned_resources_test.go
@@ -166,7 +166,7 @@ func TestEnsureIngressServiceForDataPlane(t *testing.T) {
 			expectedServiceType:      corev1.ServiceTypeLoadBalancer,
 			expectedServicePorts: []corev1.ServicePort{
 				{
-					Name:       "port-8080",
+					Name:       "http",
 					Port:       8080,
 					TargetPort: intstr.FromInt(8000),
 					Protocol:   corev1.ProtocolTCP,